+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### Not Released Yet
+
+##### Additions :tada:
+
+- Added `CesiumGeometry::Transforms::getUpAxisTransform` to get the transform that converts from one up axis to another.
+
 ### v0.40.1 - 2024-10-01
 
 ##### Fixes :wrench: